Renamee remaining pieces to pamela
