+++ shared/hooks/index.js
@@ -5,4 +5,4 @@ export * from './useWindowTitle';
 export * from './useSideDrawer.js';
 export * from './useShowSystemNamespaces';
 export * from './navigate';
-export * from './PamelaAPI/index';
+export * from './BackendAPI/index';
